+++ tests/lean_spec/subspecs/forkchoice/test_validator.py
@@ -413,7 +413,7 @@ class TestValidatorErrorHandling:
         num_validators = len(state.validators)
 
         # is_proposer_for should work (though likely return False)
-        result = large_validator.is_proposer_for(large_slot, num_validators)
+        result = large_validator.is_proposer_for(large_slot, Uint64(num_validators))
         assert isinstance(result, bool)
 
         # Attestation can be created for any validator
